+++ block_ptr/vllm_unified/triton_attn.py
@@ -87,8 +87,10 @@ def _fwd_kernel_unified(
     kv_head_idx = q_head_idx // num_queries_per_kv
 
     S = tl.load(seq_lens + batch_idx)
-    q_start = tl.load(query_start_loc + batch_idx)
-    q_end = tl.load(query_start_loc + batch_idx + 1)
+    # Cast to int32 — make_block_ptr offsets only support 32-bit (Triton constraint).
+    # query_start_loc may be int64 (PyTorch cumsum default).
+    q_start = tl.load(query_start_loc + batch_idx).to(tl.int32)
+    q_end = tl.load(query_start_loc + batch_idx + 1).to(tl.int32)
     q_len = q_end - q_start
 
     # Local q offsets within this sequence's q-token span
